Fix (Step2): View 분리
